+++ Cpp/Arrays/generateSubarrays.cpp
