+++ Cpp/Competitions/Google Kick Start/googlekickstart_Festival.cpp
