+++ Cpp/Maths/check-niven-num.cpp
