+++ Cpp/STL and Templates/DequeueSTL.cpp
