+++ Cpp/STL and Templates/TEMPLATE.cpp
