--- ROTATION/Search_in_Rotated_Sorted_Array.ctxt
@@ -1,5 +1,0 @@
-#BlueJ class context
-comment0.target=Search_in_Rotated_Sorted_Array
-comment1.params=nums\ target
-comment1.target=boolean\ search(int[],\ int)
-numComments=2
